Fix screen flicker issue before verify clinic
Had this issue were just before the verify clinic page showed you
would see the dahsboard. Couldn't figure it out until i realised
it was because Firebase gives us a cached reponse initially (and
a bunch of ther times as well). Anyway went down this rabbit hole
and eventaully had to get the cached variable so navigation can
block it. Was a little problem but took a lot to fix.