Fix city refresh issue
City field wasn't refreshing when new province was chosen and was
showing the result for the old province until you started typing
in the field. Also added functionailty so that selection of
province is not necessary to get suggestions, just typing in a
correct province in the field is enough.