Add cloud function for automatic survey export to google drive
Did a lot of work in the functions. Not entirely sure it all works,
the whole v1 and v2 thing trips me up sometimes. But as of now the
cloud functions seem to work. Also updaed past functions a bit
due to new knowledge. Didn't go with one drive cause that needs
admin permission and can't get that from school right now.